pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "red-roving-rascal"
version = "0.1.6"
description = "A lightweight API testing and evaluation toolkit"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "dboyd13" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "pydantic>=2.0",
    "httpx",
    "boto3",
]

[project.optional-dependencies]
cdk = [
    "aws-cdk-lib>=2.150.0",
    "constructs>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/dboyd13/red-roving-rascal"
Repository = "https://github.com/dboyd13/red-roving-rascal"

[tool.pytest.ini_options]
# Suppress PytestCollectionWarning for Pydantic models named TestCase/TestSuite
# that get imported into test modules.
filterwarnings = [
    "ignore::pytest.PytestCollectionWarning:.*models.*",
]

[tool.hatch.build.targets.wheel]
packages = ["src/rascal"]
